// packages/common/business/src/pages/list/type.ts
export interface User {
    readonly id: number;
    readonly lastName?: string;
    readonly firstName?: string;
    readonly birthdate?: Date;
    readonly login?: string;
    readonly isActif?: boolean;
}

export interface Account {
    readonly password?: string;
}

export type UserAccount = User & Account;

export interface UserError {
    readonly lastNameError?: string;
    readonly firstNameError?: string;
    readonly birthdateError?: string;
    readonly loginError?: string;
    readonly passwordError?: string;
}

export type PageListData = {
    users: UserAccount[];
}

export type Store<T> =  {
    getValue: () => T;
    update: (v: T) => void;
    onUpdate: (callback: (v: T) => void) => () => void;
}